+++ lib/apiarist.dart
@@ -5,3 +5,4 @@ export 'src/api_failure.dart';
 export 'src/api_endpoint.dart';
 export 'src/api_data_response.dart';
 export 'src/api_response.dart';
+export 'src/future_api_response_extension.dart';